// packages/@aws-cdk/aws-cloudwatch/lib/graph.d.ts
export interface MetricWidgetProps {
    /**
     * Title for the graph
     *
     * @default - None
     */
    readonly title?: string;
    /**
     * The region the metrics of this graph should be taken from
     *
     * @default - Current region
     */
    readonly region?: string;
    /**
     * Width of the widget, in a grid of 24 units wide
     *
     * @default 6
     */
    readonly width?: number;
    /**
     * Height of the widget
     *
     * @default - 6 for Alarm and Graph widgets.
     *   3 for single value widgets where most recent value of a metric is displayed.
     */
    readonly height?: number;
}
// ...
export declare enum LegendPosition {
    /**
     * Legend appears below the graph (default).
     */
    BOTTOM = "bottom",
    /**
     * Add shading above the annotation
     */
    RIGHT = "right",
    /**
     * Add shading below the annotation
     */
    HIDDEN = "hidden"
}